requirements.txt
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
PyYAML>=6.0.0,<7.0.0
beautifulsoup4>=4.12.0,<5.0.0
tqdm>=4.66.0,<5.0.0
openai
aisuite
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
scikit-learn>=1.3.0